agents/13-seo-content-agent/requirements.txt
strands-agents>=0.1.0
boto3>=1.34.0
strands-agents-tools>=0.1.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
